// src/infrastructure/types/types.ts
export type CommentBDModel = {
    id: string
    postId: string
    content: string
    commentatorInfo: {
        userId: string
        userLogin: string
    }
    createdAt: string
    likesInfo: {
        likesCount: number
        dislikesCount: number
        statuses: {
            // [key: string]: LikeStatus
            userId: string
            status: LikeStatus
        } []
    }
}